Bring NIC up after applying network configs
